Clinical_project/requirements.txt

pandas
numpy
scikit-learn
matplotlib
seaborn
kaggle
python-dotenv
joblib
